Remove holidays for now
